requirements.txt
Babel==2.7.0
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.3
django-cors-headers==3.2.0
django-phonenumber-field==3.0.1
djangorestframework==3.10.1
drf-yasg==1.17.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
packaging==19.2
phonenumbers==8.10.15
Pillow==6.1.0
psycopg2==2.8.3
pyparsing==2.4.5
pytz==2019.1
requests==2.22.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
six==1.13.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
whitenoise==4.1.3
